requirements.txt
annotated-types==0.6.0
camel-converter==3.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
meilisearch==0.30.0
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
typing_extensions==4.10.0
urllib3==2.2.1
